Increased Memory
Instead of removing the TRACE_ERROR function call in interrupted
charging, it was better to remove the lcd_write_strings in the debug
mode i.e. "Box ID" and "unlock Count" and leave the actual values
instead. This frees up some memory to allow the compilation of the code
with the TRACE_INFO level