Fixed LVDC infinite loop
Resolved problem where the LVDC event flag (UI_LVDC) was used instead of
the LVDC user interface state (STATE_LVDC). This previously resulted in
an infinite loop when the system entered this state. Also set the LCD to
be cleared before being turned on, to avoid the previous data being
displayed.